UA Client Hints is less than harmful (#579)
* UA Client Hints is less than harmful

After talking to Google, we learned that the additional information they
are including here is not new (as it first appeared).  On balance, that
makes this less than outright harmful.  The goal is only to replicate
information that User-Agent provides in a more structured form, with the
"high entropy" elements gated behind an Accept-CH gate.

That doesn't mean that we are fully supportive of the proposal as it
stands.

We like the idea that User-Agent might be frozen.  This is
something we have already undertaken in Firefox with some success.  This
is something we would like to see, in particular for some of the "high
entropy" elements that are added in the proposal.

We are also positive regarding a JS API that makes the information
available in a less opaque and brittle way.  In particular, giving
browsers a path to rejecting access to information where the use/privacy
trade-off is not favourable (like the WOW bit on Windows systems).  This
isn't without reservations, as some of the ambiguity present in the
string has been exploited to positive ends in the past, particularly by
browsers with less market share to avoid sites discriminating against
them.  On balance, losing that ambiguity and the associated bugs is a
net positive.

The main reservation we have regards the replication of information at
the HTTP layer into new header fields.  This is inefficient and provides
less accountability than a JS API.  The proposal has browsers send four
header fields in place of the one, which takes considerably more space.
Even with header compression, this adds non-trivial amounts of overhead
to every request.  Sites that request additional information with
Accept-CH will further inflate that.

The advantages we might realize from that design depend on the
structured information being more usable and reliable.  On the usability
side, more granular information can make caching strategies more
capable.  On the reliability side, use of structured fields means that
there should be fewer parsing irregularities and quirks.  This is a
general claim about Client Hints as a whole, just applied to User-Agent.

However, in the case of User-Agent, where the information is already
present, the trade-off is a little tougher.  Sites, and even caches,
have a means to access the information today.  It's inconvenient, but
various proprietary approaches exist to manage granular caching based on
User-Agent or Cookie; it might be worth exploring standardization of
other approaches to the same problem, if caching happens to be a major
motivating factor for the decomposition of the field.

The proponents are going to look at making the specification clearer
about its intent regarding new information.  With that understanding,
this change adjusts the position to reflect a position that is - for the
entire proposal - conflicted.

* Take Henri's suggestion

With a few edits, mostly for brevity.  This is long anyway, but it's
worth trying to edit it down.

* correct typo

Co-authored-by: Anne van Kesteren <annevk@annevk.nl>